requirements.txt
Flask==2.3.3
ariadne==0.22.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
boto3==1.34.0
botocore==1.34.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0 